Improve typings to match actual interface exposing "get" method.
Current typings don't properly expose the "get" method which is needed to monitor size on output stream resize event.  I didn't implement the full TTY interface mainly because the source module only uses this for a deprecated method and tty itself from node isn't exposed as an interface. Saw little need in defining it but it could be done for Node v.0.5 or below if someone really desired to do so.